(1/x) Move dialect files one directory up so that dialect layout will resemble the guide provided by SQLAlchemy.
